pqueue: move to a subdirectory
