+++ Makefile
@@ -16,7 +16,7 @@ ccflags-$(CONFIG_NVMEVIRT_NVM) += -DBASE_SSD=INTEL_OPTANE
 nvmev-$(CONFIG_NVMEVIRT_NVM) += simple_ftl.o
 
 ccflags-$(CONFIG_NVMEVIRT_SSD) += -DBASE_SSD=SAMSUNG_970PRO
-nvmev-$(CONFIG_NVMEVIRT_SSD) += ssd.o conv_ftl.o pqueue.o channel_model.o
+nvmev-$(CONFIG_NVMEVIRT_SSD) += ssd.o conv_ftl.o pqueue/pqueue.o channel_model.o
 
 ccflags-$(CONFIG_NVMEVIRT_ZNS) += -DBASE_SSD=ZNS_PROTOTYPE
 ccflags-$(CONFIG_NVMEVIRT_ZNS) += -Wno-implicit-fallthrough
